HealthProfessional:Define the HealthProfessional class, which contains the ID, name, basic information, and print method.
This code defines a class called HealthProfessional to represent information about health professionals. The class contains three private attributes: id, name, and basicInfo. Two constructors are provided: the default constructor initializes the default value, and the parametric constructor allows the property value to be specified when the object is created. In addition, the class defines a printDetails method to print all of the object's property values, and a getName method to get the object's name. These features enable the flexible creation and management of relevant data for health professionals.