Specialist:The code adds domain attributes and methods for printing details to represent different types of health professionals.
The Specialist class inherits from the HealthProfessional class and represents health professionals in a specific field. It includes default constructors and constructors with parameters, and can initialize all instance variables (including those in the base class). This class implements the printDetails method for printing details about health professionals, including fields and specialty types (such as "Specialist"). This class provides the basis for expanding the different types of health professionals, with the flexibility to present different medical professional information